refactor: using service for nextjs api backend
