+++ frontend/src/app/api/service/map/place/restaurant/route.ts
@@ -2,15 +2,18 @@ import type { NextRequest } from 'next/server';
 import { NextResponse } from 'next/server';
 
 import { GOOGLE_MAPS_API_KEY } from '@/libs/envValues';
-
-import getRestaurantData from './getRestaurantData';
+import mapPlaceRestaurantService from '@/service/map/place/restaurant/service';
 
 export async function POST(req: NextRequest) {
   const { latitude, longitude } = await req.json();
   const apiKey = GOOGLE_MAPS_API_KEY;
 
   try {
-    const restaurantData = await getRestaurantData(latitude, longitude, apiKey);
+    const restaurantData = await mapPlaceRestaurantService.getRestaurantData(
+      latitude,
+      longitude,
+      apiKey,
+    );
     return NextResponse.json(restaurantData);
   } catch (error) {
     console.error(error);
