+++ frontend/src/app/api/service/map/place/route.ts
@@ -2,20 +2,17 @@ import type { NextRequest } from 'next/server';
 import { NextResponse } from 'next/server';
 
 import { GOOGLE_MAPS_API_KEY } from '@/libs/envValues';
+import mapPlaceService from '@/service/map/place/service';
 import type { PlaceDetails } from '@/types/place-details';
-
-import { getPlaceDetails } from './getPlaceDetails';
-import { getPlaceId } from './getPlaceId';
-import { getPlacePhoto } from './getPlacePhoto';
 
 export async function POST(req: NextRequest) {
   const { placeName } = await req.json();
   const apiKey = GOOGLE_MAPS_API_KEY;
 
   try {
-    const placeId: string = await getPlaceId(placeName, apiKey);
-    const placeDetails: PlaceDetails = await getPlaceDetails(placeId, apiKey, 'ja');
-    const placePhoto: string = await getPlacePhoto(placeDetails.photo, apiKey);
+    const placeId: string = await mapPlaceService.getPlaceId(placeName, apiKey);
+    const placeDetails: PlaceDetails = await mapPlaceService.getPlaceDetails(placeId, apiKey, 'ja');
+    const placePhoto: string = await mapPlaceService.getPlacePhoto(placeDetails.photo, apiKey);
 
     const response: PlaceDetails = {
       placeId,
