+++ frontend/src/client/api/get-result/implement.ts
@@ -10,7 +10,7 @@ import type { ApiContext } from '@/client/ApiContext';
 import cacheClient from '@/client/service/cache/implement';
 import { API_ENDPOINT } from '@/libs/envValues';
 import { generateObjectId } from '@/libs/helper';
-import mapPlaceService from '@/service/map/place/service';
+import type { PlaceDetails } from '@/types/place-details';
 import type { Location, Recommendation } from '@/types/recommendation';
 
 // eslint-disable-next-line complexity
@@ -61,7 +61,10 @@ const generateLocation = async (activity: {
   place: string;
   description: string;
 }): Promise<Location> => {
-  const placeData = await mapPlaceService.getPlaceData(activity.place);
+  const response = await axios.post(`/api/service/map/place/`, {
+    placeName: activity.place,
+  });
+  const placeData: PlaceDetails = response.data;
 
   return {
     id: generateObjectId(),
