+++ frontend/src/service/map/place/restaurant/interface.ts
@@ -1,0 +1,5 @@
+import type { Location } from '@/types/recommendation';
+
+export interface MapPlaceRestaurantServiceInterface {
+  getRestaurantData: (lat: number, lng: number, apiKey: string) => Promise<Location[]>;
+}
